Restored custom fit method

+++ cnn-deblur/models/wgan.py
@@ -287,7 +287,7 @@ class WGAN:
             # Perform training
             for train_batch in notebook.tqdm(train_data, total=steps_per_epoch):
                 # Perform train step
-                step_result = self.train_step(tf.cast(train_batch, dtype='float32'))
+                step_result = self.train_step(train_batch)
 
                 # Collect results
                 g_losses.append(step_result['g_loss'])
@@ -332,7 +332,7 @@ class WGAN:
                 val_fake_l1_metrics = []
                 for val_batch in notebook.tqdm(validation_data, total=validation_steps):
                     # Perform eval step
-                    step_result = self.test_step(tf.cast(val_batch, dtype='float32'))
+                    step_result = self.test_step(val_batch)
 
                     # Collect results
                     val_g_losses.append(step_result['g_loss'])
@@ -602,7 +602,7 @@ class WGAN:
             # Perform training
             for batch in notebook.tqdm(train_data, total=steps_per_epoch):
                 # Perform train step
-                step_result = self.distributed_train_step(tf.cast(batch, dtype='float32'), strategy)
+                step_result = self.distributed_train_step(batch, strategy)
 
                 # Collect results
                 g_losses.append(step_result['g_loss'])
@@ -647,7 +647,7 @@ class WGAN:
                 val_fake_l1_metrics = []
                 for val_batch in notebook.tqdm(validation_data, total=validation_steps):
                     # Perform eval step
-                    step_result = self.test_step(tf.cast(val_batch, dtype='float32'))
+                    step_result = self.test_step(val_batch)
 
                     # Collect results
                     val_g_losses.append(step_result['g_loss'])
@@ -724,7 +724,7 @@ class WGAN:
         fake_l1_metrics = []
         for batch in notebook.tqdm(test_data, total=steps):
             # Perform test step
-            step_result = self.test_step(tf.cast(batch, dtype='float32'))
+            step_result = self.test_step(batch)
 
             # Collect results
             g_losses.append(step_result['g_loss'])
